Django_tutorial_practice/studybudb/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
